LayoutTests/imported/w3c/web-platform-tests/websockets/Create-invalid-urls.any-expected.txt

CONSOLE MESSAGE: line 6: Wrong url scheme for WebSocket http://localhost:8800/echo
CONSOLE MESSAGE: line 12: Wrong url scheme for WebSocket mailto:microsoft@microsoft.com
CONSOLE MESSAGE: line 18: Wrong url scheme for WebSocket about:blank
CONSOLE MESSAGE: line 24: Wrong url scheme for WebSocket http://localhost:8800/websockets/Create-invalid-urls.any.html?test
CONSOLE MESSAGE: line 30: Wrong url scheme for WebSocket http://localhost:8800/websockets/Create-invalid-urls.any.html#test

PASS Url is /echo - should throw SYNTAX_ERR 
PASS Url is a mail address - should throw SYNTAX_ERR 
PASS Url is about:blank - should throw SYNTAX_ERR 
PASS Url is ?test - should throw SYNTAX_ERR 
PASS Url is #test - should throw SYNTAX_ERR 
